Add JupyterLab link to README
BTW, is this "roadmap" project still used? It seems a bit outdated ...

I actually wanted to find this information?

Is it already decided whether JupyterLab will replace the Notebook?
If not, when will it be decided?

Will there be a feature freeze (or something similar) of the Notebook in order to concentrate work on JupyterLab?